add info and settings button on mobileheader
