+++ web/src/features/map-controls/MapControls.tsx
@@ -1,11 +1,8 @@
-import { Button } from 'components/Button';
-import { isInfoModalOpenAtom, isSettingsModalOpenAtom } from 'features/modals/modalAtoms';
-import { useAtom, useAtomValue, useSetAtom } from 'jotai';
+import { useAtom, useAtomValue } from 'jotai';
 import { useTransition } from 'react';
 import { useTranslation } from 'react-i18next';
 import { FiWind } from 'react-icons/fi';
 import { HiOutlineEyeOff, HiOutlineSun } from 'react-icons/hi';
-import { HiCog6Tooth, HiOutlineInformationCircle } from 'react-icons/hi2';
 import { MoonLoader } from 'react-spinners';
 import trackEvent from 'utils/analytics';
 import { ThemeOptions, ToggleOptions } from 'utils/constants';
@@ -19,39 +16,19 @@ import {
   windLayerAtom,
   windLayerLoadingAtom,
 } from 'utils/state/atoms';
+import { useIsMobile } from 'utils/styling';
 
 import ConsumptionProductionToggle from './ConsumptionProductionToggle';
 import { LanguageSelector } from './LanguageSelector';
 import MapButton from './MapButton';
+import MobileButtons from './MobileButtons';
 import SpatialAggregatesToggle from './SpatialAggregatesToggle';
 import ThemeSelector from './ThemeSelector';
 
 function MobileMapControls() {
-  const setIsInfoModalOpen = useSetAtom(isInfoModalOpenAtom);
-  const setIsSettingsModalOpen = useSetAtom(isSettingsModalOpenAtom);
-
-  const handleOpenInfoModal = () => setIsInfoModalOpen(true);
-  const handleOpenSettingsModal = () => setIsSettingsModalOpen(true);
-
   return (
-    <div className="absolute right-2 top-2 flex space-x-3 pt-[env(safe-area-inset-top)] sm:hidden">
-      <Button
-        size="lg"
-        type="secondary"
-        aria-label="open info modal"
-        backgroundClasses="bg-white/80 backdrop-blur-sm dark:bg-gray-800/80"
-        onClick={handleOpenInfoModal}
-        icon={<HiOutlineInformationCircle size={21} />}
-      />
-      <Button
-        size="lg"
-        type="secondary"
-        aria-label="open settings modal"
-        onClick={handleOpenSettingsModal}
-        backgroundClasses="bg-white/80 backdrop-blur-sm dark:bg-gray-800/80"
-        icon={<HiCog6Tooth size={20} />}
-        data-test-id="settings-button-mobile"
-      />
+    <div className="absolute right-0 mt-[env(safe-area-inset-top)] p-1">
+      <MobileButtons />
     </div>
   );
 }
@@ -168,10 +145,6 @@ function DesktopMapControls() {
 }
 
 export default function MapControls() {
-  return (
-    <>
-      <MobileMapControls />
-      <DesktopMapControls />
-    </>
-  );
+  const isMobile = useIsMobile();
+  return isMobile ? <MobileMapControls /> : <DesktopMapControls />;
 }
