+++ web/src/features/map-controls/MobileButtons.tsx
@@ -1,0 +1,33 @@
+import { Button } from 'components/Button';
+import { isInfoModalOpenAtom, isSettingsModalOpenAtom } from 'features/modals/modalAtoms';
+import { useSetAtom } from 'jotai';
+import { Info, Settings } from 'lucide-react';
+
+export default function MobileButtons() {
+  const setIsInfoModalOpen = useSetAtom(isInfoModalOpenAtom);
+  const setIsSettingsModalOpen = useSetAtom(isSettingsModalOpenAtom);
+
+  const handleOpenInfoModal = () => setIsInfoModalOpen(true);
+  const handleOpenSettingsModal = () => setIsSettingsModalOpen(true);
+  return (
+    <div className="mr-2 flex gap-2">
+      <Button
+        size="lg"
+        type="secondary"
+        aria-label="open info modal"
+        backgroundClasses="bg-white/80 backdrop-blur dark:bg-gray-800/80"
+        onClick={handleOpenInfoModal}
+        icon={<Info size={20} />}
+      />
+      <Button
+        size="lg"
+        type="secondary"
+        aria-label="open settings modal"
+        onClick={handleOpenSettingsModal}
+        backgroundClasses="bg-white/80 backdrop-blur dark:bg-gray-800/80"
+        icon={<Settings size={20} />}
+        data-test-id="settings-button-mobile"
+      />
+    </div>
+  );
+}
