+++ web/src/features/panels/LeftPanel.tsx
@@ -1,5 +1,6 @@
 import LoadingSpinner from 'components/LoadingSpinner';
 import Logo from 'features/header/Logo';
+import MobileButtons from 'features/map-controls/MobileButtons';
 import { useAtom } from 'jotai';
 import { lazy, Suspense } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -12,6 +13,7 @@ import {
   useParams,
   useSearchParams,
 } from 'react-router-dom';
+import { useIsMobile } from 'utils/styling';
 
 import { leftPanelOpenAtom } from './panelAtoms';
 
@@ -86,8 +88,9 @@ function CollapseButton({ isCollapsed, onCollapse }: CollapseButtonProps) {
 
 function MobileHeader() {
   return (
-    <div className="flex w-full items-center justify-between p-1 pt-[env(safe-area-inset-top)] shadow-md dark:bg-gray-900 sm:hidden">
+    <div className="mt-[env(safe-area-inset-top)] flex w-full items-center justify-between p-1 shadow-md dark:bg-gray-900">
       <Logo className="h-10 w-44 fill-black dark:fill-white" />
+      <MobileButtons />
     </div>
   );
 }
@@ -95,17 +98,18 @@ function MobileHeader() {
 function OuterPanel({ children }: { children: React.ReactNode }) {
   const [isOpen, setOpen] = useAtom(leftPanelOpenAtom);
   const location = useLocation();
+  const isMobile = useIsMobile();
 
   const onCollapse = () => setOpen(!isOpen);
 
   return (
     <aside
       data-test-id="left-panel"
-      className={`absolute left-0 top-0 z-20 h-full w-full  bg-zinc-50 shadow-xl transition-all duration-500 dark:bg-gray-900 dark:[color-scheme:dark] sm:flex sm:w-[calc(14vw_+_16rem)] ${
+      className={`absolute left-0 top-0 z-20 h-full w-full bg-zinc-50 shadow-xl transition-all duration-500 dark:bg-gray-900 dark:[color-scheme:dark] sm:flex sm:w-[calc(14vw_+_16rem)] ${
         location.pathname === '/map' ? 'hidden' : ''
       } ${isOpen ? '' : '-translate-x-full'}`}
     >
-      <MobileHeader />
+      {isMobile && <MobileHeader />}
       <section className="h-full w-full">{children}</section>
       <CollapseButton isCollapsed={!isOpen} onCollapse={onCollapse} />
     </aside>
